chore: migrates email and password fields
