+++ packages/ui/src/forms/field-types/Text/types.ts
@@ -1,6 +1,8 @@
 import type { FormFieldBase } from '../shared'
 
 export type Props = FormFieldBase & {
+  name?: string
+  path?: string
   inputRef?: React.MutableRefObject<HTMLInputElement>
   onKeyDown?: React.KeyboardEventHandler<HTMLInputElement>
 }
